backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jaydebeapi>=1.2.3
JPype1>=1.5.0
